eKYC/requirements.txt
absl-py==1.0.0
astor==0.8.1
dlib==19.20.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
importlib-metadata==1.7.0
imutils==0.5.3
Keras==2.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.22
opencv-python-4.6.0.66
opt-einsum==3.2.1
protobuf==3.20.3
PyYAML==5.3.1
six==1.15.0
tensorboard==2.13.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0


setuptools==59.5.0
torch==1.5.0
torchvision==0.6.0
scikit-learn==0.22.2
scipy==1.5.1
matplotlib==3.1.2
tqdm==4.60.0
Pillow==8.2.0
thop
torchsummary
facenet_pytorch


shapely
scikit-image
imgaug
pyclipper
lmdb
visualdl
rapidfuzz
opencv-contrib-python
cython
lxml
premailer
openpyxl
attrdict
loguru
uvicorn
python-multipart
paddlepaddle==2.5.2
mtcnn

flask
flask-cors

xlrd #for ppocr label